Create a client package, separate the api and the client
